refactor(atls-ui): move story files to stories folder
