+++ ui-parts/avatar/stories/avatar.stories.tsx
@@ -1,15 +1,15 @@
 import type { Meta }           from '@storybook/react'
 import type { StoryObj }       from '@storybook/react'
 
-import type { AvatarVariants } from './avatar/index.js'
+import type { AvatarVariants } from '../src/avatar/index.js'
 
 import React                   from 'react'
 
 import { TextTransform }       from '@atls-ui-parts/text-transform'
 
-import { AvatarFallback }      from './avatar-fallback/index.js'
-import { AvatarImage }         from './avatar-image/index.js'
-import { Avatar }              from './avatar/index.js'
+import { AvatarFallback }      from '../src/avatar-fallback/index.js'
+import { AvatarImage }         from '../src/avatar-image/index.js'
+import { Avatar }              from '../src/avatar/index.js'
 
 interface AvatarProps extends AvatarVariants {
   image: string
