+++ ui-parts/button/stories/button.stories.tsx
@@ -1,7 +1,7 @@
 import type { Meta }     from '@storybook/react'
 import type { StoryObj } from '@storybook/react'
 
-import { Button }        from './button.component.js'
+import { Button }        from '../src/button.component.js'
 
 const meta: Meta<typeof Button> = {
   title: 'Components/Button',
