+++ ui-parts/button/stories/ghost-button.stories.tsx
