+++ ui-parts/switch/stories/switch.stories.tsx
@@ -1,14 +1,14 @@
 import type { Meta }        from '@storybook/react'
 import type { StoryObj }    from '@storybook/react'
 
-import type { SwitchProps } from './switch.interfaces.js'
+import type { SwitchProps } from '../src/switch.interfaces.js'
 
 import React                from 'react'
 
 import { Column }           from '@atls-ui-parts/layout'
 import { Layout }           from '@atls-ui-parts/layout'
 
-import { Switch }           from './switch.component.js'
+import { Switch }           from '../src/switch.component.js'
 
 const meta: Meta<SwitchProps> = {
   title: 'Components/Switch',
